refactoring ProxyService and ToDoService.
Add the user queue and fix sync with the azure service.

+++ todoclient/ToDoClient/Services/ProxyService.cs
@@ -38,35 +38,68 @@ namespace todoclient.Services
         /// The url for a todo's deletion.
         /// </summary>
         private const string DeleteUrl = "ToDos/{0}";
-        public static int UserId { get; set; }
-        
+        public static Queue<int> listOfUsersId = new Queue<int>();
+        private Thread ProxyThread { get; set; }
+
         private readonly HttpClient _httpClient;
+        private static ProxyService instance;
+
+        private static object syncRoot = new Object();
+
         public ProxyService()
         {
             _httpClient = new HttpClient();
             _itemRepository = new ItemRepository();
             _httpClient.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
+         
         }
 
-        public void UpdateAzureService()
+        public static ProxyService GetInstance()
         {
-            while (true)
+            if (instance == null)
             {
-                var dataAsString = _httpClient.GetStringAsync(string.Format(_serviceApiUrl + GetAllUrl, UserId)).Result;
-                var userViewItems = JsonConvert.DeserializeObject<IList<ToDoItemViewModel>>(dataAsString);
-                var itemsIdsFromAzureService = userViewItems.Select(i => i.ToDoId.ToString());
-                foreach (var id in itemsIdsFromAzureService)
+                lock (syncRoot)
                 {
-                    _httpClient.DeleteAsync(string.Format(_serviceApiUrl + DeleteUrl, id))
-                            .Result.EnsureSuccessStatusCode();
+                    if (instance == null)
+                    { 
+                        instance = new ProxyService();
+                        instance.StartProxy();
+                    }
                 }
-                var listOfItemsFromDb = _itemRepository.GetItems(UserId).Where(i => i.UserId == UserId).Select(i => i.ToViewModel());
-                foreach (var item in listOfItemsFromDb)
-                {
-                    ThreadPool.QueueUserWorkItem(t => _httpClient.PostAsJsonAsync(_serviceApiUrl + CreateUrl, item)
-                        .Result.EnsureSuccessStatusCode());
-                }
+            }
+            return instance;
+        }
+        public void StartProxy()
+        {
+            ProxyThread = new Thread(UpdateAzureService) { IsBackground = true };
+            ProxyThread.Start();
+        }
+        private void UpdateAzureService()
+        {
+            while(true)
+            {
+                if(listOfUsersId.Count != 0)
+                { 
+                    int userId = listOfUsersId.Dequeue();
+                    var dataAsString = _httpClient.GetStringAsync(string.Format(_serviceApiUrl + GetAllUrl, userId)).Result;
+                    var userViewItems = JsonConvert.DeserializeObject<IList<ToDoItemViewModel>>(dataAsString);
+                    var itemsIdsFromAzureService = userViewItems.Select(i => i.ToDoId.ToString());
 
+                    foreach (var id in itemsIdsFromAzureService)
+                    {
+                        _httpClient.DeleteAsync(string.Format(_serviceApiUrl + DeleteUrl, id))
+                            .Result.EnsureSuccessStatusCode();
+                    }
+
+                    IEnumerable<ToDoItemViewModel> listOfItemsFromDb =
+                        _itemRepository.GetItems(userId).Where(i => i.UserId == userId).Select(i => i.ToViewModel());
+
+                    foreach (var item in listOfItemsFromDb)
+                    {
+                        _httpClient.PostAsJsonAsync(_serviceApiUrl + CreateUrl, item)
+                            .Result.EnsureSuccessStatusCode();
+                    }
+                }
             }
         }
     }
